more updates, sim works now

+++ sw/airborne/subsystems/gps.h
@@ -65,8 +65,8 @@ struct GpsState {
   int32_t hmsl;                  ///< height above mean sea level in mm
   struct EcefCoor_i ecef_vel;    ///< speed ECEF in cm/s
   struct NedCoor_i ned_vel;      ///< speed NED in cm/s
-  int32_t gspeed;                ///< norm of 2d ground speed in cm/s
-  int32_t speed_3d;              ///< norm of 3d speed in cm/s
+  int16_t gspeed;                ///< norm of 2d ground speed in cm/s
+  int16_t speed_3d;              ///< norm of 3d speed in cm/s
   int32_t course;                ///< GPS heading in rad*1e7
   uint32_t pacc;                 ///< position accuracy
   uint32_t sacc;                 ///< speed accuracy
